==> 00-competition/eda/team/01-About Dataset.ipynb <==
# 1.라이브러리

import warnings
warnings.filterwarnings('ignore')

import os
import pandas as pd

# 2.데이터 경로 지정 및 파일 읽기

data_path = '../data'
train_path = f'{data_path}/train.csv'
test_path  = f'{data_path}/test.csv'

print(os.listdir(data_path))

train_data = pd.read_csv(train_path)
test_data = pd.read_csv(test_path)
print('Train data shape : ', train_data.shape, 'Test data shape : ', test_data.shape)

sampled_data = train_data.sample(n=20)
sampled_data.to_csv('sampled_data.csv', index=False)

# 3.컬럼과 결측치 확인

train_data.info()

for idx, column in enumerate(train_data.columns, start=1):
    print(f"{idx:>03} : {column}")

train_data.info()

# EDA에 앞서 결측치를 확인해보겠습니다.
train_data.isnull().sum()

display(train_data["해제사유발생일"].unique())
display(train_data["해제사유발생일"].value_counts())

display(train_data["등기신청일자"].unique())
display(train_data["등기신청일자"].value_counts())

display(train_data["거래유형"].unique())
display(train_data["거래유형"].value_counts())

display(train_data["중개사소재지"].unique())
display(train_data["중개사소재지"].value_counts())

display(train_data["k-단지분류(아파트,주상복합등등)"].unique())
display(train_data["k-단지분류(아파트,주상복합등등)"].value_counts())

display(train_data["k-세대타입(분양형태)"].unique())
display(train_data["k-세대타입(분양형태)"].value_counts())

display(train_data["k-관리방식"].unique())
display(train_data["k-관리방식"].value_counts())

display(train_data["k-복도유형"].unique())
display(train_data["k-복도유형"].value_counts())

display(train_data["k-난방방식"].unique())
display(train_data["k-난방방식"].value_counts())

display(train_data["k-건설사(시공사)"].unique())
display(train_data["k-건설사(시공사)"].value_counts())

display(train_data["k-시행사"].unique())
display(train_data["k-시행사"].value_counts())

display(train_data["경비비관리형태"].unique())
display(train_data["경비비관리형태"].value_counts())

display(train_data["세대전기계약방법"].unique())
display(train_data["세대전기계약방법"].value_counts())

display(train_data["청소비관리형태"].unique())
display(train_data["청소비관리형태"].value_counts())

test_data.isnull().sum()
