==> 00-competition/eda/team/1_X_Y_preprocess.ipynb <==
# 좌표 X, Y 결측치 처리
- '시군구', '번지'를 합쳐 하나의 feature '시군구 번지' 생성
- 크롤링을 통해 좌표를 찾는다.
- '시군구 번지'에 맞게 결측치를 처리한다.

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
from tqdm.notebook import tqdm
import time
import pickle
import gc

from selenium import webdriver
from selenium.webdriver.common.by import By
from selenium.webdriver.support.ui import WebDriverWait
from selenium.webdriver.support import expected_conditions as EC

from sklearn.preprocessing import OneHotEncoder, MinMaxScaler, StandardScaler
from sklearn.model_selection import train_test_split

# 맥에서 글씨 깨짐 방지
from matplotlib import rc
rc('font', family='AppleGothic')
plt.rcParams['axes.unicode_minus'] = False

## DataLoad

df = pd.read_csv('../data/train_test_data.csv')
df

df.shape

## X, Y 좌표가 없는 데이터만 추출

tmp = df[['시군구', '번지', '아파트명', '좌표X', '좌표Y']]
x_y_nan = tmp[tmp['좌표X'].isna()]

x_y_nan

x_y_nan.info()

a = x_y_nan['시군구']
b = x_y_nan['번지']
x_y_nan['시군구 번지'] = a+' '+b

x_y_nan

print(x_y_nan.shape)
print(len(x_y_nan['시군구 번지'].unique()))
locates = x_y_nan['시군구 번지'].unique()

## 크롤링

url = 'https://www.findlatlng.org/'
driver = webdriver.Chrome()
driver.get(url)
time.sleep(20)

# '힐스테이트 서초 젠트리스': (127.05347005393612, 37.45628519999)
re = dict()
cnt = 1

tbar = tqdm(locates)
for locate in tbar:
    driver.find_element(By.CLASS_NAME, 'form-control').clear()
    driver.find_element(By.CLASS_NAME, 'form-control').click()
    driver.find_element(By.CLASS_NAME, 'form-control').send_keys(locate)
    driver.find_element(By.CLASS_NAME, 'btn').click()
    time.sleep(1)
    container = driver.find_elements(By.CLASS_NAME, 'container-fluid')
    try:
        data = container[1].text.split('\n')[1].split(' / ')
        X = data[1][16:]
        Y = data[0][15:]
        re[locate] = (float(X), float(Y))
    except:
        continue

    if cnt % 500 == 0:
        finded_x_y_df = pd.DataFrame(re).transpose()
        finded_x_y_df.columns = ['좌표X', '좌표Y']
        finded_x_y_df.to_csv(f'./data/finded_x_y_{cnt//500}.csv')

        del re
        del finded_x_y_df
        gc.collect()
        re = dict()
    
    cnt += 1

finded_x_y_df = pd.DataFrame(re).transpose()
finded_x_y_df.columns = ['좌표X', '좌표Y']
finded_x_y_df.to_csv(f'./data/finded_x_y_last.csv')

# 찾은 X, Y 좌표 데이터
X_Y_df = pd.concat([pd.read_csv(f'../data/x_y/finded_x_y_{i}.csv') for i in range(1, 18)], axis=0)

X_Y_df.columns = ['시군구 번지', '좌표X', '좌표Y']
X_Y_df.reset_index(drop=True)

## 결측치 채워넣기

# 시군구, 번지가 합쳐진 하나의 변수 생성

a = df['시군구']
b = df['번지']
df['시군구 번지'] = a+' '+b

df

# '시군구 번지'에 맞춰 X, Y 를 매핑

X_mapping = pd.Series({locate: X for locate, X in zip(X_Y_df['시군구 번지'], X_Y_df['좌표X'])})
Y_mapping = pd.Series({locate: Y for locate, Y in zip(X_Y_df['시군구 번지'], X_Y_df['좌표Y'])})

X_mapping

# 원본 데이터에 적용
df['좌표X'] = df['좌표X'].fillna(df['시군구 번지'].map(X_mapping))
df['좌표Y'] = df['좌표Y'].fillna(df['시군구 번지'].map(Y_mapping))

# 힐스테이트 서초 젠트리스는 번지가 없기에 따로 작업

df[df['좌표X'].isna()]['아파트명'].unique()

# 직접 찾은 좌표 적용

hill_X_mapping = pd.Series({'힐스테이트 서초 젠트리스': 127.05340799855254})
hill_Y_mapping = pd.Series({'힐스테이트 서초 젠트리스': 37.457157542923255})

df['좌표X'] = df['좌표X'].fillna(df['아파트명'].map(hill_X_mapping))
df['좌표Y'] = df['좌표Y'].fillna(df['아파트명'].map(hill_Y_mapping))

print(df['좌표X'].isna().sum())
print(df['좌표Y'].isna().sum())

df.to_csv('../data/preprocessed/X_Y_preprocessed.csv', index=False)
