==> 00-competition/eda/team/12_district_EDA.ipynb <==
# 행정구역 구분

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

# 맥에서 글씨 깨짐 방지
from matplotlib import rc
rc('font', family='AppleGothic')
plt.rcParams['axes.unicode_minus'] = False

df = pd.read_csv('../data/preprocessed/11_const_rank.csv')

df['시군구'].unique()

df[['시', '구', '동']] = df['시군구'].str.split(' ', expand=True)

df.drop(['시'], axis=1, inplace=True)

len(df['구'].unique())

district_li = df['구'].unique().tolist()
df_split = []

for district in district_li:
    df_split.append(df[df['구'] == district])

df_split[1]

# 분할된 데이터를 다시 합쳐준다.
# 인덱스에 맞게 정렬

tmp = pd.concat([splited_df for splited_df in df_split]).sort_index()

df.to_csv('../data/preprocessed/12_district_split.csv', index=False)
